Put dehydration in separate package
